refactor(MobileMsg): moved file into /components folder and update import in App
